# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "SaveLibraryRequest": {
        "properties": {
          "audioBitrateKbps": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "audioClippingGateEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "audioLoudnessGateEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "audioTargetCodec": {
            "type": [
              "null",
              "string"
            ]
          },
          "autoEnqueueEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "autoEnqueueWindowEnd": {
            "type": [
              "null",
              "string"
            ]
          },
          "autoEnqueueWindowStart": {
            "type": [
              "null",
              "string"
            ]
          },
          "autoReplace": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "clipVmafEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "contentTune": {
            "type": [
              "null",
              "string"
            ]
          },
          "cropBlackBars": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "downmixToStereo": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "durationTolerancePercent": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "enabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "encoderPreset": {
            "type": [
              "null",
              "string"
            ]
          },
          "excludeHardLinkedFiles": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "excludePaths": {
            "type": [
              "null",
              "string"
            ]
          },
          "hdrHandling": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageDownscaleMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageDownscaleValue": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "imageMetadataGateEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "imageQuality": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "imageQualityGateEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "keepAudioLanguages": {
            "type": [
              "null",
              "string"
            ]
          },
          "keepSubtitleLanguages": {
            "type": [
              "null",
              "string"
            ]
          },
          "maxBitrateKbps": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "maxFrameRate": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "maxHeight": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "maxLoudnessDriftLufs": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "maxTruePeakDbtp": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "mediaType": {
            "type": [
              "null",
              "string"
            ]
          },
          "minBitrateKbps": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "minFileSizeBytes": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "minVmafCatastrophicMin": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "minVmafHarmonicMean": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "minVmafMin": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "minimumImageSsim": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          },
          "moveOnComplete": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "moveOverwrite": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "optimiseDolbyVision": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "path": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "qualityCrf": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "reencodeLossyAudio": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "reencodeLossyImages": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "reencodeSameCodecAboveBytes": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "requireAudioRetained": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "requireSizeReduction": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "requireSubtitlesRetained": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "ruleProfile": {
            "type": [
              "null",
              "string"
            ]
          },
          "skipEfficientSources": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "skipSourceCodecs": {
            "type": [
              "null",
              "string"
            ]
          },
          "strongerAdaptiveQuantisation": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "targetContainer": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetFolder": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetImageFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "targetVideoCodec": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoAudioBitrateKbps": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "videoAudioCodec": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoDownscaleHeight": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "videoQualityStrategy": {
            "type": [
              "null",
              "string"
            ]
          },
          "vmafFrameSubsample": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "vmafQualityGateEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "workPlacement": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "path",
          "mediaType",
          "ruleProfile",
          "enabled",
          "priority",
          "minFileSizeBytes",
          "maxHeight",
          "videoDownscaleHeight",
          "maxFrameRate",
          "cropBlackBars",
          "reencodeSameCodecAboveBytes",
          "skipEfficientSources",
          "targetVideoCodec",
          "targetContainer",
          "hdrHandling",
          "optimiseDolbyVision",
          "excludePaths",
          "excludeHardLinkedFiles",
          "skipSourceCodecs",
          "contentTune",
          "maxBitrateKbps",
          "minBitrateKbps",
          "strongerAdaptiveQuantisation",
          "qualityCrf",
          "encoderPreset",
          "audioTargetCodec",
          "audioBitrateKbps",
          "videoAudioCodec",
          "videoAudioBitrateKbps",
          "downmixToStereo",
          "keepAudioLanguages",
          "keepSubtitleLanguages",
          "reencodeLossyAudio",
          "targetImageFormat",
          "imageQuality",
          "reencodeLossyImages",
          "imageDownscaleMode",
          "imageDownscaleValue",
          "moveOnComplete",
          "targetFolder",
          "moveOverwrite",
          "minVmafHarmonicMean",
          "minVmafMin",
          "vmafQualityGateEnabled",
          "minVmafCatastrophicMin",
          "clipVmafEnabled",
          "vmafFrameSubsample",
          "autoEnqueueEnabled",
          "autoEnqueueWindowStart",
          "autoEnqueueWindowEnd",
          "autoReplace"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Optimisarr \u2014 safe, verified FFmpeg transcoding for self-hosted media libraries. This contract is still pre-1.0 and may change between releases.",
    "title": "Optimisarr API",
    "version": "0.2.13.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/libraries": {
      "get": {
        "operationId": "ListLibraries",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      },
      "post": {
        "operationId": "CreateLibrary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      }
    },
    "/api/libraries/scan": {
      "post": {
        "operationId": "ScanAllLibraries",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      }
    },
    "/api/libraries/{id}": {
      "delete": {
        "operationId": "DeleteLibrary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      },
      "put": {
        "operationId": "UpdateLibrary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      }
    },
    "/api/libraries/{id}/access": {
      "get": {
        "operationId": "CheckLibraryAccess",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      }
    },
...
    "/api/libraries/{id}/scan": {
      "post": {
        "operationId": "ScanLibrary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Admin token required when OPTIMISARR_ADMIN_TOKEN is set on the server."
          }
        },
        "tags": [
          "Libraries"
        ]
      }
    },
...
  },
...
}